MB-70710 Pause and resume a chat
A new statement pattern is introduced to
pause and persist a chat session.

Statement: "PAUSE AI CHAT [WITH <options>]"

The chat session's information is encoded
in a document and inserted into the QUERY_METADATA
bucket's system collection. This is similar to
what is done already for cbo histogram documents.

The key is created using the chatId which is
already unique, the key is chatId prefixed with "aichat::".
This is done keeping in mind that a new system keyspace would
provided to list or delete the paused chat.

The encoded information includes:
* the cluster user that created the chat.
* the keyspace list provided associated for the chat.
* the prompt when the pause was requested.

Pausing may also involve summarizing of the
current chat conversation when needed.

To reverse a pause, the inverse process is introduced
to resume a chat session.

Statement: "RESUME AI CHAT [WITH <options>]"

The encoded chat document is read and decoded into
a chat entry that is added to the chat cache on the
node where the resume was requested.

To handle a race between simultaneous resume
requests compare and swap check is used via
update operation on the chat document. Each resume
request updates the chat document with the
claiming node's address as well as claim timestamp.
The winning updater claims the chat document, decodes it
and adds to it's chat cache then deletes the chat document.

Incase of a claimer crashing before deleting the claimed
chat document. A new resume request first checks
the claim timestamp, if it is more than 2 mins past it.
The claim is invalidated and the resume request initiates a fresh
claim attempt.

Monitoring of paused chats can be done via the
system:natural_chats keyspace or by looking up
QUERY_METADATA's system collection. Something like,
SELECT decode_base64(chat) as chat, meta().id as chatid
FROM QUERY_METADATA._system._query
WHERE META().id LIKE "aichat::%";

Reviewed-on: https://review.couchbase.org/c/query/+/241903
Tested-by: Gaurav J <gaurav.jayaraj@couchbase.com>

+++ server/http/admin_accounting_endpoint.go
@@ -3257,6 +3257,10 @@ func doNaturalChats(endpoint *HttpEndpoint, w http.ResponseWriter, req *http.Req
 					return nil, errors.NewNaturalLanguageRequestError(errors.E_NL_CHAT_WRONG_USER)
 				}
 				ce.Lock()
+				if ce.Removed || ce.Paused {
+					ce.Unlock()
+					return true, nil
+				}
 				natural.DeleteConversation(chatId)
 				ce.Removed = true
 				ce.Unlock()
